First draft with working Execution and ExecutionLog for a progress function

+++ ArduinoMateV2/app/src/main/java/com/gmail/raducaz/arduinomate/db/dao/DeviceDao.java
@@ -30,8 +30,8 @@ public interface DeviceDao {
     void insertAll(List<DeviceEntity> devices);
 
     @Query("select * from devices where id = :deviceId")
-    LiveData<DeviceEntity> loadDevice(int deviceId);
+    LiveData<DeviceEntity> loadDevice(long deviceId);
 
     @Query("select * from devices where id = :deviceId")
-    DeviceEntity loadDeviceSync(int deviceId);
+    DeviceEntity loadDeviceSync(long deviceId);
 }
